pyproject.toml
[project]
name = "buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "hoangndst", email = "hoangndst@gmail.com" },
]
dependencies = [
    "aiohttp>=3.11.0",
    "discord-py>=2.5.2",
    "google-genai>=1.16.1",
    "tabulate>=0.9.0",
]
license = { text = "MIT" }

[dependency-groups]
dev = [
    "black>=25.1.0",
]
